refactor(providers): extract OpenAIChatCompletionsProvider abstract base — both openai-compatible + litellm now extend it
Lifts the entire orchestration pipeline (executeStream, runStreamLoop,
streamOneStep, executeToolBatch, resolveModelName, buildDelegatingModel,
getAvailableModels default impl) out of the two concrete providers into
a shared abstract base class `OpenAIChatCompletionsProvider`.

After this:

- `openaiCompatible.ts`  1018 → 166 lines (-852, -84%)
- `litellm.ts`            850 → 426 lines (-424, -50%)
- `openaiChatCompletionsBase.ts`  new (901 lines, all shared)

Concrete subclasses now declare only:

- Constructor (resolve env/credential config)
- `getProviderName()` (abstract)
- `getDefaultModel()`  (abstract)
- `formatProviderError()` (abstract)
- `getFallbackModelName()` / `getFallbackModels()` (optional)
- `adjustBuildBodyOptions()` (optional — LiteLLM uses this for the
  Gemini 2.5 maxTokens skip)
- `onStreamStart()` (optional — LiteLLM uses this for the OTel span
  wrap with cost via `calculateCost`)
- `getAvailableModels()` override (only when caching / non-default path,
  e.g. LiteLLM's 10-min cache + LITELLM_FALLBACK_MODELS env override)

LiteLLM-specific public surface (`embed`, `embedMany`, `fetchModelsFromAPI`,
the model cache statics) stays inline in `litellm.ts`. Same with the
`OpenAICompatibleProvider`-specific seven-model fallback list.

`OpenAICompatStreamLifecycleListeners` type lives in
`src/lib/types/openaiCompatible.ts` per CLAUDE.md Rule 2 + Rule 12.

Verification:

- Typecheck: 0 errors
- Lint: 0 errors / 15 pre-existing warnings
- test:bugfixes: 78/79 PASS — the 1 fail (`openai-compatible.executeStream
  HTTP failure does not produce unhandledRejection`) is a pre-existing flake
  on `release` HEAD, confirmed via `git stash && pnpm run test:bugfixes`.
  An unrelated previous test leaks an unhandledRejection about
  `/tmp/neurolink-test-creds.json` which this test captures.
- test:client: 13/13 PASS
- Live CLI `generate` + `stream` against LiteLLM (model `open-large`)
  both work end-to-end via the new base class.

Sets up the remaining 13 OpenAI-shaped providers (groq, perplexity, xai,
fireworks, togetherAi, cohere, cloudflare, huggingFace, llamaCpp, lmStudio,
openAI, deepseek, nvidiaNim) for follow-up migrations — each one becomes
a ~150-200 line subclass once the hooks for per-call credentials and
`withClientStreamSpan` OTel patterns are added to the base (separate PRs).
